PythonApplication1/requirements.txt
flask
flask-cors
sentence-transformers
torch
pdfminer.six
python-docx
